requirements.txt
pytesseract
pymupdf
Pillow
PySide6
